refactor(schema): split discovery.rs into discovery/ submodules (#1131)
Behavior-preserving module split per epic #1116. The 2,328-line
schema/discovery.rs is decomposed by responsibility into a discovery/
directory:

- model.rs:     shared data model (config, SchemaInfo, TypeInfo, validation
                result types)
- inference.rs: TypeInferenceEngine (recursive type inference)
- validator.rs: SchemaValidator + confidence/validation helpers
- exporter.rs:  SchemaExporter (CQL/JSON export, comparison report)
- analysis.rs:  engine data-analysis phases (header extract, sampling, UDT/
                collection/index discovery, build_schema_info)
- mod.rs:       SchemaDiscoveryEngine orchestration, DiscoveryContext, and
                pub use re-exports keeping the public surface unchanged

Tests follow their code: the generate_comparison_report test moves into
exporter.rs; the other 5 stay in mod.rs (need private access). Test count
preserved (6). No semantic changes.
